Avoid undefined behavior using CFlatData in CScript serialization.
https://github.com/bitcoin/bitcoin/commit/fa126effc2a03e22708960344e62fc21259deb23

`&vch[vch.size()]` and even `&vch[0]` on vectors can cause assertion
errors with VC in debug mode. This is the problem mentioned in https://github.com/bitcoin/bitcoin/pull/4239 .
The deeper problem with this is that we rely on undefined behavior.

- Add `begin_ptr` and `end_ptr` functions that get the beginning and end
  pointer of vector in a reliable way that copes with empty vectors and
  doesn't reference outside the vector
(see https://stackoverflow.com/questions/1339470/how-to-get-the-address-of-the-stdvector-buffer-start-most-elegantly/1339767#1339767).
- Add a convenience constructor to CFlatData that wraps a vector.

`begin_ptr` and `end_ptr` added as separate functions as I imagine
they will be useful in more places.